rag_project/requirements.txt
pymongo==4.6.1
qdrant-client==1.7.0
beautifulsoup4==4.12.2
requests==2.31.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
pandas==2.1.3
youtube-transcript-api==0.6.1
ollama
clearml
gradio
langchain-community
langchain-ollama